Revert the use of 3-dimensional arrays in .get_p_clust().
(Instead, for the augmented-data approach, it probably makes sense to
define `mu` as an augmented-rows matrix and to use attributes in the
family object to convert this augmented-rows matrix to a 3-dimensional
array only where necessary. That would avoid changing lots of the
projpred code to handle 3-dimensional arrays for `mu`. Other changes are
still required, though.)